Can now request for a dorm and cancel that request

+++ src/app/[user]/(user)/[book]/api/cancelDormRequest/route.js
@@ -1,8 +1,8 @@
 import pool from "@/dbConn"
 
 export async function POST(request) {
-    const res = await request.json()
-    console.log("Shopping Cart form data", res)
+    // const res = await request.json()
+    // console.log("Shopping Cart form data", res)
 
     const poolPromise = pool.promise()
     const conn = await poolPromise.getConnection()
@@ -11,14 +11,11 @@ export async function POST(request) {
     let userID = searchParams.get('userID')
     let bookID = searchParams.get('bookID')
 
-    const [rows, fields] = await conn.execute(
-      "DELETE FROM shopping_cart WHERE userID=? AND bookID=?", 
-      [parseInt(userID), parseInt(bookID)]
-    )
-    console.log("shopping cart rows", rows)
+    const [rows, fields] = await conn.execute("DELETE FROM shopping_cart WHERE userID=? AND bookID=?", [parseInt(userID), parseInt(bookID)])
+    // const [ret] = rows
+    console.log("rowsInCancelDormsRequest", rows)
     poolPromise.releaseConnection(conn)
 
-    console.log(rows)
-   
+
     return Response.json(rows)
 }
